add motor_control :: motor_a_set_power() and motor_b_set_power()

+++ motor_control.h
@@ -9,8 +9,8 @@
 #include "hard.h"
 
 /************************** PUBLIC FUNCTION **************************/
-void motor_a_set_power(int16_t power);
-void motor_b_set_power(int16_t power);
+void motor_a_set_power(int8_t power);
+void motor_b_set_power(int8_t power);
 void motors_stop();
 void motor_init();
